src/content_creation_agency/requirements.txt
agency_swarm
openai
python-dotenv
tavily-python
nltk
pytrends
google-api-python-client
textblob
pandas
